Starting Detail Activity on click

+++ app/src/main/java/com/udacity/popularmovies/themoviedb/api/MovieInfo.java
@@ -9,7 +9,7 @@ public class MovieInfo
     public String poster_path;
     public String backdrop_path;
     public String overview;
-    public int vote_average;
+    public float vote_average;
     public String original_title;
     public String release_date;
